:white_check_mark: test publish endpoint
